Rename root test directory to test-tools
