+++ packages/web-components/@sfx/sayt/tests/setup.ts
@@ -1,1 +1,1 @@
-import '../../../../../test/setup';
+import '../../../../../test-tools/setup';
