+++ packages/web-components/@sfx/sayt/tests/utils.ts
@@ -1,1 +1,1 @@
-export * from '../../../../../test/utils';
+export * from '../../../../../test-tools/utils';
